requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-genai==1.31.0
